common: add redis proxy
To handle request/response inside gRPC stream i need to store data with a TTL.
Since for now i will only use very small part of the entire Redis feature catalog.
I choose to implement a small redis proxy to have the feature without adding an entire redis instance & framework to the project.
Goal is to mimic the redis API to easily replace the proxy by a real redis in the future if we need it.

I tought i need to work with redis hash table but a simple key is enough.
This commit purpose is only to save my work with hash table, it could be interesting in the future.