chat: implem. chat stream sequence number
sequence number is used when a chat event need a response.
Example: we need to know if the private message has been received
by the recipient.

record: Basic redis replication using HashMap
Useful to store a chat event request, waiting for the answer.
With a builtin time to live for the stored data.

db_cli: chat implementation was deliberately left in it's raw phase,
and dates from the beginning of the chat development.
It still works, and could be useful to quickly test the server chat implem.
The best to do, would be to adapt and use the model/view/controller implem. for the client.
But it's not the priority for now.