client: precompute collider map to prepare pathfinding
This new collider maps are now used for the collision detection.

The current 'w/a/s/d' deplacement will be removed and replaced
by a mouse clic and a pathfinding algorithm to move the entities.
Therefore the current detection collision will also be removed.

Assumptions:
first frame of the collider layer is the same accross all next frames
(Its the case for now with the maps we have, but we should keep this in mind for future one)