project: rename map coordinate into cell
In the project we have two type of coordinates:
The coord which locate the map in the world
The coord which locate entities on the map
'map/Map/MapCoord' was the naming used for entities location within a map, this was misleading.
From now to locate an entities on a map, 'cell' will be used.
Since one map is a tilemap, similar to a grid, and a grid is composed of cells.

This commit is the first part of the coordinates renaming.
The second part will rename 'world/World/WorldCoord'