client: fix chat window rendering, limit window msg number
I try to keep 3 WindowMessage display but it also
mean keeping an offset (see msg_raw in render function)
for each entities, in this commit we have only one offset
for all entities.
This cause an offset bug when someone else is talking.
But it's works correctly for one entities talking.
Useful to check the feature (keep 3 messages window) to test
but should be fixed if adopt.